+++ src/plastimatch/plm_image_convert.cxx
@@ -78,7 +78,7 @@ plm_image_convert_itk_to_gpuit_float (Plm_image* pli, T img)
     typename ImageType::DirectionType dc = img->GetDirection();
 
     /* Copy header & allocate data for gpuit float */
-    int dim[3];
+    size_t dim[3];
     float offset[3];
     float spacing[3];
     float direction_cosines[9];
@@ -300,7 +300,7 @@ plm_image_convert_itk_uchar_vec_to_gpuit_uchar_vec (Plm_image* pli,
     UCharVecImageType::SpacingType sp = itk_img->GetSpacing();
     UCharVecImageType::SizeType sz = rg.GetSize();
     UCharVecImageType::DirectionType itk_dc = itk_img->GetDirection();
-    int dim[3];
+    size_t dim[3];
     float offset[3];
     float spacing[3];
     float direction_cosines[9];
